requirements.txt
numpy
pandas
matplotlib
librosa
torch
torchsummary
tqdm
streamlit
loguru
streamlit-player
blinker
click
protobuf
python-dateutil
streamlit_option_menu
lyricsgenius
openai-whisper

